+++ Model/Api/V2.php
@@ -386,28 +386,29 @@ class V2 implements V2Interface
             );
         }
     }
-        /**
-         * @inheritDoc
-         */
-        public function sendConfig(string[] $config): void
-        {
-            $uri = $this->config->getGatewayURL() . self::SEZZLE_SEND_CONFIG_ENDPOINT;
-            try {
-                $transferO = $this->transferFactory->create(array_merge([
-                    '__method' => Client::HTTP_POST,
-                    '__uri' => $uri
-                ], $config));
-                $response = $this->client->placeRequest($transferO);
 
-                if (!empty($response) || $response["status_code"] !==     Response::STATUS_CODE_204) {
-                    throw new Exception(__("Invalid status code: " . $response["status_code"]));
-                }
-            } catch (Exception $e) {
-                $this->helper->logSezzleActions($e->getMessage());
-                throw new LocalizedException(
-                    __('Gateway send config details error: %1', $e->getMessage())
-                );
+    /**
+     * @inheritDoc
+     */
+    public function sendConfig(array $config): void
+    {
+        $uri = $this->config->getGatewayURL() . self::SEZZLE_SEND_CONFIG_ENDPOINT;
+        try {
+            $transferO = $this->transferFactory->create(array_merge([
+                '__method' => Client::HTTP_POST,
+                '__uri' => $uri
+            ], $config));
+            $response = $this->client->placeRequest($transferO);
+
+            if (!empty($response) && $response["status_code"] !== Response::STATUS_CODE_204) {
+                throw new Exception(__("Invalid status code: " . $response["status_code"]));
             }
+        } catch (Exception $e) {
+            $this->helper->logSezzleActions($e->getMessage());
+            throw new LocalizedException(
+                __('Gateway send config details error: %1', $e->getMessage())
+            );
         }
+    }
 
 }